Revert "small bug fix, for type of category features"
This reverts commit e768249b2074c3c3889ad9bb2950ba837b35f8ab.

+++ smogn/over_sampling.py
@@ -581,9 +581,8 @@ def over_sampling(
 
         #for x in code_list:
         #    data_new.iloc[:, j] = data_new.iloc[:, j].replace(x, cat_list[x])
-        # The categorical feature should be Integer not decimal
         mapping = mapping_dict[j]
-        data_new.iloc[:, j] = data_new.iloc[:, j].map(mapping).astype('int64')
+        data_new.iloc[:, j] = data_new.iloc[:, j].map(mapping)
 
     ## reintroduce constant features previously removed
     if len(feat_const) > 0:
